Rework index commands
