+++ src/main/java/frc/team2412/robot/commands/index/IndexMoveBallToFirstSensor.java
@@ -3,28 +3,25 @@ package frc.team2412.robot.commands.index;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.team2412.robot.subsystem.IndexSubsystem;
 
-public class IndexIngestMotorOutCommand extends CommandBase {
+public class IndexMoveBallToFirstSensor extends CommandBase {
     private final IndexSubsystem subsystem;
 
-    public IndexIngestMotorOutCommand(IndexSubsystem subsystem) {
+    public IndexMoveBallToFirstSensor(IndexSubsystem subsystem) {
         this.subsystem = subsystem;
         addRequirements(subsystem);
     }
 
     @Override
     public void initialize() {
-        subsystem.ingestMotorOut();
-    }
-
-    @Override
-    public void execute() {
-        if (!subsystem.ingestSensorHasBallIn()) {
-            subsystem.ingestMotorStop();
-        }
+        subsystem.ingestMotorIn();
     }
 
     @Override
     public boolean isFinished() {
-        return !subsystem.isIngestMotorOn();
+        if (subsystem.ingestSensorHasBallIn()) {
+            subsystem.ingestMotorStop();
+            return true;
+        }
+        return false;
     }
 }
